+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixes a potential issue with rate limiting.
+
 ## [v15.6.6](https://github.com/lexicalunit/spellbot/releases/tag/v15.6.6) - 2025-08-24
 
 ### Changed